some fix for the push

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: egiraldi <egiraldi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 01:49:26 by egiraldi          #+#    #+#             */
-/*   Updated: 2021/11/08 01:59:05 by egiraldi         ###   ########lyon.fr   */
+/*   Updated: 2021/12/02 11:56:24 by egiraldi         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_bzero(void *s, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
