+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: egiraldi <egiraldi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 01:48:45 by egiraldi          #+#    #+#             */
-/*   Updated: 2021/11/08 01:48:46 by egiraldi         ###   ########lyon.fr   */
+/*   Updated: 2021/12/02 13:05:19 by egiraldi         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *str)
 {
 	char	*copy;
-	int		len;
+	size_t	len;
 
 	len = ft_strlen(str);
 	copy = (char *)malloc(sizeof(char) * (len + 1));
