+++ ft_strrev.c
@@ -6,7 +6,7 @@
 /*   By: egiraldi <egiraldi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 01:48:27 by egiraldi          #+#    #+#             */
-/*   Updated: 2021/11/08 01:48:28 by egiraldi         ###   ########lyon.fr   */
+/*   Updated: 2021/12/02 13:07:22 by egiraldi         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strrev(char *str)
 {
-	int		i;
-	int		size;
+	size_t	i;
+	size_t	size;
 	char	temp;
 
 	i = 0;
